(PC-2113): Move count_users_having_booked() to user repository
