+++ tests/repository/user_queries_test.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta, MINYEAR
 from models import ImportStatus
 from models import PcObject
 from repository.user_queries import get_all_users_wallet_balances, find_by_civility, \
-    find_most_recent_beneficiary_creation_date, count_activated_users
+    find_most_recent_beneficiary_creation_date, count_activated_users, count_users_having_booked
 from tests.conftest import clean_database
 from tests.test_utils import create_user, create_offerer, create_venue, create_offer_with_thing_product, create_deposit, \
     create_stock, create_booking, create_beneficiary_import
@@ -260,6 +260,59 @@ class CountActivatedUsersTest:
 
         # Then
         assert number_of_active_users == 0
+
+
+class CountUsersHavingBookedTest:
+    @clean_database
+    def test_returns_one_when_user_with_one_cancelled_and_one_non_cancelled_bookings(self, app):
+        # Given
+        user_having_booked = create_user()
+        offerer = create_offerer()
+        venue = create_venue(offerer)
+        offer1 = create_offer_with_thing_product(venue)
+        offer2 = create_offer_with_thing_product(venue)
+        stock1 = create_stock(offer=offer1, price=0)
+        stock2 = create_stock(offer=offer2, price=0)
+        booking1 = create_booking(user_having_booked, stock1, is_cancelled=False)
+        booking2 = create_booking(user_having_booked, stock2, is_cancelled=True)
+        PcObject.save(booking1, booking2)
+
+        # When
+        number_of_users_having_booked = count_users_having_booked()
+
+        # Then
+        assert number_of_users_having_booked == 1
+
+    @clean_database
+    def test_returns_two_when_two_users_with_cancelled_bookings(self, app):
+        # Given
+        user_having_booked1 = create_user()
+        user_having_booked2 = create_user(email='test1@email.com')
+        offerer = create_offerer()
+        venue = create_venue(offerer)
+        offer = create_offer_with_thing_product(venue)
+        stock = create_stock(offer=offer, price=0)
+        booking1 = create_booking(user_having_booked1, stock, is_cancelled=True)
+        booking2 = create_booking(user_having_booked2, stock, is_cancelled=True)
+        PcObject.save(booking1, booking2)
+
+        # When
+        number_of_users_having_booked = count_users_having_booked()
+
+        # Then
+        assert number_of_users_having_booked == 2
+
+    @clean_database
+    def test_returns_zero_when_no_user_with_booking(self, app):
+        # Given
+        user = create_user()
+        PcObject.save(user)
+
+        # When
+        number_of_users_having_booked = count_users_having_booked()
+
+        # Then
+        assert number_of_users_having_booked == 0
 
 
 def _create_balances_for_user2(stock3, user2, venue):
